add required methods to TestCafeCore API
